Reset Password Feature
users are now able to reset their passwords. users will input their email address which will be verified in the system, if the email exists in our database then a reset password link is sent and the user is able to securely change their password for up to one hour before that link is expired.

must add reset_tokens table to database code for that is :
create table reset_tokens(
	id INT AUTO_INCREMENT PRIMARY KEY,
	email VARCHAR(100) NOT NULL UNIQUE,
    token VARCHAR(50) NOT NULL UNIQUE,
    expiration DATETIME NOT NULL,
    FOREIGN KEY (email) REFERENCES users(email)
);